Moved Logic from save image page to home_page
